refactor: separate codebase - copy
